Define RAID level for LV and use it to determine required PV space
We already support various segment types for LVs to some extent. However, we
need to get a better picture of how much space on such LVs' PVs is required. We
already have a code for that, so let's just use it.

This unfortunately requires the LVPVSpec to have read-write attributes/fields
and thus it cannot be a namedtuple. We should probably come up with some
"read-write namedtuple" thing for cases like this in the future.