Fix pylint problems with the new default namedtuple stuff.
(1) Something's going on with pylinting util_test.py leading to
tracebacks that end line this:

  File "/usr/lib/python3.4/site-packages/astroid/builder.py", line 39, in parse
    return compile(string, "<string>", 'exec', PyCF_ONLY_AST)
  File "<string>", line 2
    class 1(tuple):

No one wants to wade into that, so just disable pylint on that one file
for the moment.

(2) Don't worry about TheDefaultNamedTuple not having an __init__.
That's just one pylint can't figure out.

(3) Get rid of an unnecessary lambda.