Do not try to get LVM cache's size from stats for inactive LV
If the cached LV is not active, we cannot determine the size of its cache from
the stats because these are only available for activated LVs. However, if the LV
and cache exist, we examined them on devicetree population and got the
information about the size at that point. Thus we can used the stored value if
the LV is no longer active.

Also, keep the logic of when we should be able to get stats in one place.