Account for LVM RAID metadata
LVM creates small internal LVs for RAID LVs that hold the necessary metadata. We
need to account for that in order to be able to do calculations of PV/VG free
space etc. This requires us to do some of the calculations in a more granular
manner -- separating data and metadata parts.