Add PEP8 compliance checking to 'check' target.
Moves 'pylint' to its own target, adds a 'pep8' target, and makes
the 'check' target empty with 'check' and 'pep8' as prerequisites.
This allows the user to run either pylint or pep8 separately or
together.

It also makes it possible to plumb a user-specified PYTHONPATH
into the 'pylint' target.

We are currently ignoring errors for overly-long lines, hence the
'--ignore E501'.