Track information about free space in existing PVs
We need that information in order to do checks when adding more LVs and users
need this information to decide about where to place their LVs.

Let's not bother with existing LVs allocations for now. We can just ignore those
and only care about newly added (non-existing) LVs which we need to place
somewhere.