Replace the 'copies' property of LVs with something more accurate
The word "copies" is accurate together with mirror/RAID1 RAID, but it's
misleading with other RAID levels LVM supports. The property doesn't seem to be
accessed anywhere outside the class so let's just replace it with a private
property with a more accurate name.

Also give incomplete/inaccurate information if we have incomplete/inaccurate
information instead of erroring out.