Fix the twisted logic of lv.metadata_size
lv.size reports the size of the LV not the space occupied in the VG, that's what
data_vg_space_used is for. Under the same logic lv.metadata_size should report
the size of the metadata space LV has available leaving
lv.metadata_vg_space_used for reporting how much space from the VG the metadata
part(s) of the LV take.

If the LV exists, we should just go through the internal metadata LVs and sum
their sizes because that's the actual/real value.

Also document the property.

Please note that no changes are needed outside of these two properties because
they are already used properly and this just fixes the values such places in
code calculate with (like metadata_size passed to blockdev.thpoolcreate() or
calculation of the pmspare LV's size).