Account for LVM cache space in PVs
Now that we keep track of available space in the PVs we need to take into
account LVM caches because those specify PVs and thus we need to make sure that
they really fit in somewhere. Also the users need to know how much space they
still have available in their PVs if they add a cache to their LV(s).