Account for a pmspare LV creation or growth when determining the cache size
When an LVM cache is created, an internal metadata LV is created for it. And for
LVM that also means that a special pmspare LV with a size greater or equal to
the size of the metadata LV has to exist (and thus may be created) in the same
VG. We don't want to bother user code with these calculations and thus we should
subtract this space from the requested cache's size.