Remove some extraneous methods related to ignored devices.
This is just a bunch of code that has been mostly obsolete for a
long time.

is_ignored doesn't actually consult the list of ignored/exclusive
disks -- instead, it just adds some additional cases in which we
might want to ignore a device. I think this information is better
stored elsewhere, if only to reduce the number of methods with the
word "ignore" in the name.

Things ignored because of is_ignored:

  1. /dev/ram*

    This is already handled by blivet.udev.__is_blacklisted_blockdev.

  2. loop devices with no backing file

    This is a good idea, and is preserved by an accompanying patch
    to populator.handlers.LoopDevicePopulator's match method.

Things not ignored because of is_ignored (a no-op, really):

  1. md container arrays
  2. fwraid arrays whose name was mangled by mdadm because of
     HOMEHOST.
  3. mapped disk images
  4. live media images
  5. loop devices with a backing file

The only place we call add_ignored_disk is right after is_ignored
returns True for a device. It is not appropriate in any of the
cases flagged by is_ignored to add the device to the lvm device
blacklist.