Always use setup.py sdist to generate source tarballs.
Use sdist for both the "local" and "archive" make targets so that the
only difference between the two is that "archive" is built from a
specific git tag.

Include the .mo files in the sdist, and extend the sdist command from
setup.py to build the .mo files before creating the archive.

Be less weird about making the source tarball in "make local".