Add a decorator for guarding methods and properties with a property
Sometimes we need to prevent a method from being called or property from being
evaluated if some conditition is not satisfied. The most common case is to check
value of some property of the given object. For example, it only makes sense to
call the lv.merge() method if 'lv' is a snapshot. Things like this can be
facilitated by a decorator that makes sure 'lv' is snapshot for example by
checking its 'is_snapshot' property. The code then looks like this:

@requires_property("is_snapshot")
def merge(self):
    # the actual merge goes here

On an LV that is not a snapshot a ValueError is raised.

Of course, the decorator can be made even more generic if it turns out in the
future we need it to be.