Rewrite the code for representing LVs
Having specific classes for various types of LVs is nice because Python makes
sure the right code is executed based on the type (class) of the LV. However, a
big drawback of this is that it is not possible to change the type (class) of
the object once it is created. This makes things like 'lvconvert' or even thin
snapshot creation and activation impossible to represent.

Thus we need to represent all LVs with single class making a lot of decisions on
our own. However, that doesn't mean we have to put all the code into a single
class. By utilizing some nice tricks, we can easily put type-specific code into
type-specific mixin classes and make the final/ultimate class inherit from
them. That way the type of an LV is more a dynamic property than a
once-and-for-always-given thing.