Add facilities for reporting exceptions that occur in threads.
Callers can register a callback which blivet can use to report
exceptions that occur in event handling threads. If they do not
register such a callback, the following will happen: 1) The exception
and traceback will be logged using the ERROR facility. 2) The next time
a method that has been wrapped using the SynchronizedMeta metaclass is
called in the main thread, the exception will be re-raised.