Adapt our code to changes in the representation of LVs
All LVs are now instances of a single class. The type is determined by the
properties of the object like its segment type, whether it has a parent LV
(internal LVs) or an origin (snapshots). Thus we need to make sure the right
values are passed in all the places we create LV objects.