Add a test for the to-be-used @type_specific decorator
When objects of various (sub)types need to be represented by a single class
(e.g. because their (sub)types can change in time), it's nice to implement the
(sub)type specific code in mixin classes the single resulting class inherits
from. With thatthere are two (major) options how to implement the methods of
such "ultimate" class:

1) Have gigantic ifs at the beginning of the methods making sure the right code
   is executed for the particular object and its (sub)type.
2) Implement and use the @type_specific decorator that makes sure the method
   of the right mixin class is called.

This adds an example implementation of 2) included in the test. A generic
@type_specific decorator would be even more complex and harder to use which is
why the decorator is tailored for the particular class and lives under its
namespace.