Add decorator for guarding methods with external dependency checks
Some methods can only be called if their external dependencies are
available. Otherwise they should fail either silently (logging a warning and
returning None) or hard (raising an exception). Also, we should be able to
evaluate the state and availability of the external dependecies always or just
once during the first call.