Allow for pre-release builds.
The versioning will follow the conventions set forth in
http://docs.openstack.org/developer/pbr/semver.html, which
incorporates pre-release versioning set forth in PEP-0440
into semantic versioning.

To do a pre-release build of, say 2.0.0 alpha 1, the upstream
version is set to '2.0.0.a1'. In the rpm spec file we set the
version to '2.0.0', the numeric portion of release from '1' to
'0.1', and then define prerelease as '.a1'. This yields a
release of '0.1.a1'. When bumping the release for a bugfix
build, you increment the '.1' in the release, yielding '0.2'
as the numeric portion and a full release value of '0.2.a1'.