Do not pretend thin pools take more space in their VGs
When creating a thin pool we are leaving quite a big padding for it to be grown
later. However, that's not a space taken from the VG, it's just a space we are
reserving which can be used for any other purpose. Thus we shouldn't pretend
it's an extra space taken by the thin pool.