Fix blivet constructor fs support check
Calls to get_default_filesystem_type check to see if the filesystem is
supported by the installed operating system. The boolean flag that
indicated support was actually returning an instance of the property
object which, when cast to a boolean type, is always True.

Converted to call the actual boolean property and made the error handler
return False instead of None for consistency with the property.

Related: rhbz#1242666