Fix blivet constructor fs support check (#1256249)
Users who run blivet tests without root permissions would get test
failures if they did not have the ext4 kernel modules loaded and did not
have permission to load kernel modules.

Fixed the guard to check the boolean value of the property rather than
checking for the existence of the pointer. Added xfs as a default
filesystem as it is another commonly used filesystem.

Related: rhbz#1256249