Get rid of blivet.blivet's dependence on blivet.osinstall.FSSet
All of the FSSet stuff pertains to installation, so move everything
from Blivet to InstallerStorage subclass. The methods and properties
moved are:
    file_system_free_space
    root_device (prop)
    update_ksdata
    _update_custom_storage_ksdata
These now only exist in InstallerStorage.

This also creates the "mountpoints" property in blivet.devicetree

This property fits pretty nicely in conjunction with other properties
living inside of the devicetree. It also simplifies things in
blivet.osinstall.FSSet.