Protect the live device's parent (#1172342)
When an iso9660 is written to a USB the child partitions are excluded
from the devicetree by addUdevPartitionDevice. This means it never gets
to the point where it can protect the parent devices.

This adds a check to addUdevPartitionDevice that will protect the whole
disk if any of the partitions are the live backing device.