Protect the live device's parent (#1172342)
When an iso9660 is written to a USB the child partitions are excluded
from the devicetree by addUdevPartitionDevice. This means it never gets
to the point where it can protect the parent devices.

This adds a second check to see if the liveBackingDevice name starts
with any of the other device names. If so then the device is set as
protected and added to the protectedDevNames list.