Correctly recognize internal RAID LVs
If an internal LV is a RAID LV, its lv_attr is really interesting. Usually (bug
in LVM?) it has 'r' as the first letter so we need to allow that and then do
some magic to correctly determine the role(s) of the LV.

Using 'lvs -o+lv_role' would be much better for this, but I only learnt about it
today and using it would require more changes (incl. some in libblockdev).