Isolate some boot device operations in blivet.partitioning.
Two pieces of bootloader logic are a bit isolated now, thanks to an
optional kwarg which can be passed into do_partitioning and
allocate_partitions (boot_disk).

The default value for this kwarg is None, since a Blivet instance is
not going to have this defined (or need to worry about it). However,
if it is defined, e.g. in an InstallerStorage instance, one should
take care to pass this value in if/when a call to these functions is
made.

The logic which has been isolated handles sorting the boot disk to be
at the beginning of the disk list. Another piece of logic sets the boot
flag on the boot device.

This also specifies a Blivet spec to use in the Mock constructor for
partitioning tests, otherwise they fail by trying to run a piece of
code specific to InstallerStorage.