Move some more installer-specific methods and values into blivet.osinstall.
This moves all of the default stub values defined in blivet.__init__ to
blivet.osinstall, as well as some methods, so:

* all of the constants
* enable_installer_mode
* get_sysroot
* get_target_physical_root
* set_sysroot

blivet.blivet does rely upon calling short_product_name (one of the pieces
moved into osinstall) as well as sysroot and storage root. To account for
that, created some properties in the Blivet class with a default value.
These are rw properties, so they can be changed if need be.

This also moves update_from_anaconda_flags() out of blivet.flags and
into blivet.osinstall, since it's another installer-specific piece. Some
flags which were set in update_from_anaconda_flags() are now set in
enable_installer_mode() -- these are flags that were not actually
updated based on anaconda flags.