Create flags.auto_dev_updates
This is intended to be a flag set to False by default in blivet, since
(forced) automatic updates of a device's info or status should not be
necessary. In some circumstances (e.g. in the installer), a situation
may arise where it is necessary, so this flag is set to True in those
instances.