Get rid of flags.installer_mode
Lots of stuff happens here to reach the end goal of no installer_mode:

* shutdown() is mv'ed from blivet.blivet to blivet.osinstall; as this
only runs if you're doing an installation, that seemed appropriate.

* every flags.installer_mode check was removed from blivet.osinstall;
since you are going to be installing if you use anything in
blivet.osinstall, checking for installer_mode is redundant.

* every flags.installer_mode check was removed from blivet.udev;
now there is a blivet.device_name_blacklist list which is empty by
default; if there are any members, the list is iterated over and
matching devices are blacklisted as expected.

* blivet.blivet now relies on some localized logic to determine
whether or not to run additional code, instead of checking
flags.installer_mode.

* installer_mode was removed from blivet.flags, and is no longer set
in enable_installer_mode() method.

* selinux test was altered to compensate for the loss of
flags.installer_mode