Ensure biosboot shows up in kickstart (#1242666)
Added support to detect the biosboot partition.

The biosboot partition, unlike other bootloaders, is not considered a stage1
device and is not mountable therefore it never actually showed up. Detection
of the biosboot partition required a special case that does not impact the
internal bootloader device properties which would have wreaked havoc
elsewhere in the codebase.

Resolves: rhbz#1242666