Support disk tags as disk specification for new partitions.
To specify disk set using tags when instantiating PartitionDevice,
pass an iterable containing one or more disk tag values via the
keyword argument 'disk_tags'. If multiple tags are passed, the
value is interpreted as "any of these tags". If a non-empty
parents list is passed, the tags will be ignored. Similarly, if
a non-empty list of tags is passed, an empty parents list is
ignored in favor of the tags.