Require BlockDev 2.0 in the gi.require_version() call
We now have libblockdev-2.0. Unfortunately, the useless function
gi.require_version() is stupid and doesn't support anything like ">= 1.0", which
is what we really need. So unless we want to add more useless code trying both
"1.0" and "2.0", we are doomed to fail on systems that have libblockdev-1.x
installed even if that's just fine for us. To prevent this let's also bump the
requires in the spec file.

Resolves: rhbz#1395791