tasks/fsmkfs: Add arguments for setting UUID
These arguments are only relevant if we want to set a UUID during
creation of file systems.

Right now this doesn't implement any real functionality and only
implements the interface and the options for various file systems.

The reason I'm using get_uuid_args instead of just uuid_options is that
for XFS, we need to use ["-m", "uuid=something"], so we don't have a
single flag with a plain argument.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>