formats/fs: Add functionality for checking UUIDs
This is more or less modelled the same way as tasks.fslabeling,
consisting of classes that currently only contain a method for verifying
whether an UUID is correct for a given file system or not.

I'm using _uuidfs within the FS class (and subclasses) even though it
sounds a bit odd, mainly because it is consistent with _labelfs and also
because it avoids clashes with future _uuid attributes.

For the _check_rfc4122_uuid() class method, I'm using string processing
rather than uuid.UUID(), because first of all uuid.UUID() is much more
complicated and it also allows to omit the dashes while the file system
utilities may not allow this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>