formats/fs: Implement setting UUID during mkfs
We already have a uuid keyword argument, which might get passed if the
format already exists, but if it exists on a new format, we can safely
assume that the user wants the format to have a specific UUID and we use
that instead of letting the mkfs utility generate a random UUID.

The implementation for _create() deviates a bit from the scheme we had
with labels so far, where we have a labeling() method which returns True
if we can set a label either during or after the creation of a file
system.

In our case we now have a can_assign_uuid(), which *only* returns True
if we can assign an UUID during creation. We will have another method
soon which checks whether we can assign an UUID *after* creation.

Also, using a similar naming scheme to labeling(), we would have ended
up with something like uuiding() or any other weird name. IMHO setting
this to can_assign_uuid() also has the advantage that the name in itself
is more descriptive about what this method does.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>