tasks: Implement setting UUID after FS creation
Some file systems don't have utilities for setting the UUID during
creation, so we need to fall back to doing it afterwards in
_post_create().

This introduces mlabel (for FAT) and tune2fs (for ext2/3/4) as new
applications required in tasks/availability.

The "mlabel" tool is required to set serials for FAT file systems,
because dosfstools seems to only allow setting the volume ID during
creation and doesn't have a command to change it for existing file
systems.

"fatlabel" only can set labels, but not UUIDs but "mlabel" can.

Unfortunately, I haven't found a way to set HFS+ UUIDs after creation
using the hfsprogs package, so I'm omitting HFS+ here for now.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Edited-by: Vratislav Podzimek <vpodzime@redhat.com>