tests: Add a series of tests for setting UUIDs
While we already had pure tests for just verifying whether an UUID is
correct, let's now add the impure stuff that actually uses the various
file system tools to write the UUIDs for real.

Note that the fsuuid.py file does not use the same scheme than
fslabeling.py, even though it might look similar.

For UUIDs we assume here, that we can _always_ assign a new UUID after
the file system has been created. Of course, this is not yet true for
HFS+, because I haven't yet found a way to set the UUID afterwards.

So the distinction between SetUUIDWithMkFs and SetUUIDAfterMkFs is that
with the former we assume that the mkfs utility has a way to set the
UUID, even though we have a _post_create() method that sets the UUID if
that's not the case.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>