formats/swap: Support setting UUID
This is quite similar to what we have for file systems, but doesn't rely
on the fsuuid/fswriteuuid tasks except for the _check_rfc4122_uuid()
method from fsuuid.

Other than that the implementation is the same.

We can't assign the UUID for a swap file system at a later point, so we
only implement this during creation.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>