Make padding smaller for existing thin pools (#1432012)
Because when creating them we could have not know the metadata
size and relied on the padding to provide enough space for
it. Then for an existing thin pool, we know the metadata size and
if we just sum it with the padding we can easily get over the
VG's (free space) size resulting in a VG reporting negative free
space.