Reserve space in a VG instead of padding thin pools on autopart
Padding thin pools brings many complications and issues. For
example, what should we do with existing thin pools? What if
somebody really wants to create a thin pool of size not leaving
enough space for the padding in the VG? And more.

Nevertheless, we want to leave some space in the VG if doing
thinp auto partitioning because that's a good practice. However,
we already have an existing mechanism for that -- the VG can
reserve some space in itself (the size or percentage of which can
even be defined by user). So let's just extend and use this
mechanism to make sure there's some free space in the VG left for
the thin pool to grow when doing autopartitioning. If users give
us all the numbers, we should just do as they say, they probably
know what they are doing.

(ported commit 788d7d0d8fcaa3c830bdcdc24482e1b7ea815920)

Resolves: rhbz#1131247