Check if new device's size is != 0 when comparing it to maxSize
Setting size to 0 happens as part of the operations in
DeviceFactory (when removing a device). And when that happens,
the calculated maxSize can be negative (as an interim
state). However, setting device's size to 0 in order to make sure
its container can omit the space for it should be always allowed.

Related: rhbz#1131247