Try to mount and unmount an XFS FS when writing UUID
xfs_admin refuses to change the UUID of an unclean FS.
Unfortunately, this is quite a common case for file systems
created as part of creating a snaphot of the underlying block
device if it's mounted at the time of the snapshot
creation. Fortunately, the solution is as simple as just mounting
and unmounting the file system because that cleans it up.