Look the disk up for a partition by name not sys_name
udev.device_get_partition_disk() returns the device name as given
by device_get_name() not sys_name. So when trying to find a match
we need to compare it to the same value for devices we iterate
over.

Also don't call resolve_devspec() on the returned value,
udev.device_get_partition_disk() is already doing it.

Resolves: rhbz#1445302