Always mount & unmount an XFS file system when writing new UUID
This is needed when the file system is not clean (e.g. if it was
created as a snapshot of some other mounted file system) which is
quite a common case. And since 'xfs_admin' exits with code 0 even
when it fails to set the UUID [1], we cannot do a second/fallback
try. The first one has to have the best chance to succeed, hence
this patch.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1450423