Refer to self.container as self.vg in the LVMFactory
Because that's just what it really is. And it makes things like
'self.container.pe_size' less confusing because no other thing
than a VG has the 'pe_size' attribute. Of course all the code in
the LVMFactory is specific to LVM where self.container is always
a VG (or None). Even in the LVMThinPFactory for which it also
makes things more explicit by mitigating the risk of confusing
self.pool and self.container (because that factory actually has
two containers).