pylint: ignore "arguments-differ" warning for "do_tasks" method
This actually isn't a false positive because "do_tasks" really
changes the number of arguments when overriding the abstract
method but it would be necessary to change huge piece of the
tasks code to fix this. So just mark this as false positive
and fix it when (and if) replacing/removing the tasks.