Only consider old and new device sizes when not growing to max
If a container (VG) is requested to grow to its maximum size,
there's no point in taking into account the size of the requested
device and a potential old device when calculating the total
space it needs. It will always be the maximum its parents (or
available disks) allow.

This simplifies the code quite a bit and it allows us to move it
next to the other code dealing with the same case.

Also, we can get rid of an obsolete misleading comment.