Make sure an LV is deactivated before removal
Before and LV is removed, we may need to set its ancestors (VG
and its PVs) up which may trigger auto-activation of the LV
itself. And since we are supposed to always remove a deactivated
LV so we have to make sure the LV is deactivated before we try to
remove it.

Resolves: rhbz#1456821