Use libblockdev runtime dependency checks (#1501249)
By default libblockdev will fail to load a plugin during its
initialization if one (or more) of the plugin runtime dependencies
is missing. For example if the "swaplabel" utility is missing,
loading of the swap module will fail.
This commit disables dependency checks during init -- in the same
situation swap plugin will load and only calling "swap_set_label"
would fail (because the utility it uses is missing).
Specification of all required "technologies" and "modes" is added
to the "tasks" module so checking for external tools availability
still checks if libblockdev supports everything we need.