Move the thpool reserve calculations to LVMFactory
If a plain LV is being added to a VG that contains a thin pool,
extra space must be allocated for the VG so that the reserve for
thin pool can be grown accordingly to how the VG is grown to
include the new plain LV (because the reserve is based on the
VG's size).

The above goes through the mechanisms of the LVMFactory not
LVMThinPFactory because a plan LV is being added. Thus the code
that makes sure the above described adaptations are done needs to
be in the LVMFactory. Since LVMThinPFactory inherits from the
LVMFactory and uses the parent class' calculations to get the
base size it later modifies, the thin pool reserve is taken into
account when adding/modifying both plain and thin LVs.