Recalculate thpool's metadata size on resize in LVMThinPFactory
If the size of the pool is changed, so must be its metadata space
size. Otherwise the initial size of the pool determines the size
of its metadata space which may be totally different than an
optimal value for the end size of the pool.

For example, if a small thin LV is added first, say 1GiB, and a
huge one is added later, say 100GiB, the end size of the pool is
101 GiB. But unless the metadata space size is adapted when the
second/big thin LV is added, the pool ends up with minimum
metadata space allocated and can easily soon run out of metadata
space.

Resolves: rhbz#1490136