Add, test and use a new method to get size with reserve
Getting a size which has some X% reserve for the original size to
grow is a surprisingly complicated thing. It's quite
counter-intuitive that e.g. "120 GiB" does not (!) have a 20%
reserve for "100 GiB" to grow (because "120 * 0.2 != 20"). It's
e.g. "10 GiB" that has 20 % space reserve for "8 GiB" to grow.

That's why we should keep this black, hard-core maths in a
separate well-tested method that is used in places where we need
to do this calculation.