Add initial support for DM Integrity "format"
This format is used only for standalone integrity devices (so not
for LUKS 2 with integrity). We don't support this format right now,
even libblockdev doesn't support working with integrity devices
(it is not possible to open/setup the format). This only adds the
format to blivet so it isn't confused by it if someone creates
this setup manually.