Make raid_level a property of an LV object
With cached LVs determining the RAID level is not trivial. It's thus better to
have a property for an LVs RAID level instead of just an attribute. However, it
is not allowed to change the RAID level of an existing LV object, so the value
can be cached in an internal attribute.