Use the size info of internal LVs when getting space usage for existing LVs
We now build the whole tree of LVs and their internal LVs for existing
LVM setups. Making use of this complete information for calculating
space taken from VGs by LVs makes these calculations a lot more
precise than our magic formulas.

The reason for this is that the setups can be much more complicated
than what is properly modeled by our code, especially in case of
combinations of LVM Cache/ThinP and LVM RAID. E.g. a cached thin pool
using different RAID levels of data and metadata for both the cache
and thin pool is just too much. But the crazy tree of LVs and their
internal LVs is recognized and processed correctly even in such crazy
cases allowing us to do precise and correct calculations.

This is important for support of cases where a very complex LVM setup
has to be properly recognized, but doesn't have to be modified which
is e.g. the case of system reinstallation.