Allow duplicate UUIDs until/unless a by-uuid query occurs.
There is a new flag to control this behavior: allow_inconsistent_config.
When False, DuplicateUUIDError will be raised when trying to add any
subsequent device whose UUID matches one already in the tree. When True,
the duplicate UUIDs will be allowed to coexist until/unless a call to
DeviceTree.get_device_by_uuid is made with the duplicate UUID as its
argument. In this case, DuplicateUUIDError will be raised.