General protection against tracebacks during populate.
The goal here is to absolutely minimize tracebacks that occur when
populating the devicetree. There are mechanisms in place that will limit
the supported functionality for managing such devices, but there will be
no tracebacks until/unless the user tries to manage them in ways that
are precluded by the detection issues previously encountered.